fix: extract AbstractFormatter and AbstractHandler into separate files to fix composer 2 namespace issue
